+++ src/test/java/com/example/demo/Utility/StreamCalculatorUtilsTest.java
@@ -6,19 +6,19 @@ import org.junit.jupiter.api.Test;
 import java.util.Arrays;
 import java.util.List;
 
-class UtilsTest {
+class StreamCalculatorUtilsTest {
 
     @Test
     public void validInput() {
         String input = "123321";
-        boolean result = Utils.checkInput(input);
+        boolean result = StreamCalculatorUtils.checkInput(input);
         Assertions.assertTrue(result);
     }
 
     @Test
     public void validParse() {
         String input = "124";
-        List<Integer> result = Utils.strToIntList(input);
+        List<Integer> result = StreamCalculatorUtils.strToIntList(input);
         List<Integer> expected = Arrays.asList(1, 2, 4);
         Assertions.assertEquals(result, expected);
     }
@@ -26,21 +26,21 @@ class UtilsTest {
     @Test
     public void invalidInput() {
         String input = "123as321";
-        boolean result = Utils.checkInput(input);
+        boolean result = StreamCalculatorUtils.checkInput(input);
         Assertions.assertFalse(result);
     }
 
     @Test
     public void emptyInput() {
         String input = "";
-        boolean result = Utils.checkInput(input);
+        boolean result = StreamCalculatorUtils.checkInput(input);
         Assertions.assertFalse(result);
     }
 
     @Test
     public void nullInput() {
         String input = null;
-        boolean result = Utils.checkInput(input);
+        boolean result = StreamCalculatorUtils.checkInput(input);
         Assertions.assertFalse(result);
     }
 
@@ -48,7 +48,7 @@ class UtilsTest {
     public void invalidParse() {
         String input = "121dfg4";
         Exception exception = Assertions.assertThrows(IllegalArgumentException.class,
-                () -> Utils.strToIntList(input));
+                () -> StreamCalculatorUtils.strToIntList(input));
         String expectedMessage = "Argument is null or has illegal type";
         String actualMessage = exception.getMessage();
         Assertions.assertTrue(actualMessage.contains(expectedMessage));
@@ -58,7 +58,7 @@ class UtilsTest {
     public void emptyParse() {
         String input = "";
         Exception exception = Assertions.assertThrows(IllegalArgumentException.class,
-                () -> Utils.strToIntList(input));
+                () -> StreamCalculatorUtils.strToIntList(input));
         String expectedMessage = "Argument is null or has illegal type";
         String actualMessage = exception.getMessage();
         Assertions.assertTrue(actualMessage.contains(expectedMessage));
@@ -67,7 +67,7 @@ class UtilsTest {
     public void nullParse() {
         String input = null;
         Exception exception = Assertions.assertThrows(IllegalArgumentException.class,
-                () -> Utils.strToIntList(input));
+                () -> StreamCalculatorUtils.strToIntList(input));
         String expectedMessage = "Argument is null or has illegal type";
         String actualMessage = exception.getMessage();
         Assertions.assertTrue(actualMessage.contains(expectedMessage));
